# Use single @output_buffer when rendering