--- a/lib/view_component/base.rb
+++ b/lib/view_component/base.rb
@@ -10,6 +10,53 @@ require "view_component/slotable"
 require "view_component/slotable_v2"
 require "view_component/with_content_helper"
 
+class ActionView::Base
+  attr_reader :children
+  attr_accessor :parent
+
+  def _prepare_context
+    @children = []
+    @parent = nil
+
+    super
+  end
+
+  def set_output_buffer(buf)
+    @output_buffer = buf
+  end
+
+  def output_buffer=(buf)
+    if parent.nil?
+      set_output_buffer_for_children(buf)
+    else
+      top_level_parent.set_output_buffer_for_children(buf)
+    end
+  end
+
+  def top_level_parent
+    return if parent.nil?
+    return @_top_level_parent if defined?(@_top_level_parent)
+
+    @_top_level_parent = begin
+      top_level = parent
+
+      while top_level.parent
+        top_level = top_level.parent
+      end
+
+      top_level
+    end
+  end
+
+  def set_output_buffer_for_children(buf)
+    set_output_buffer(buf)
+
+    children.each do |child|
+      child.set_output_buffer(buf)
+    end
+  end
+end
+
 module ViewComponent
   class Base < ActionView::Base
     include ActiveSupport::Configurable
@@ -92,10 +139,17 @@ module ViewComponent
       @__vc_content_evaluated = false
       @__vc_render_in_block = block
 
+      self.parent = view_context
+
+      top_level_parent = @_top_level_parent = parent&.top_level_parent || parent
+      top_level_parent.output_buffer ||= ActionView::OutputBuffer.new
+      @output_buffer = top_level_parent.output_buffer
+      top_level_parent.children << self
+
       before_render
 
       if render?
-        render_template_for(@__vc_variant).to_s + _output_postamble
+        capture { render_template_for(@__vc_variant).to_s + _output_postamble }
       else
         ""
       end