--- a/performance/benchmark.rb
+++ b/performance/benchmark.rb
@@ -15,14 +15,19 @@ class BenchmarksController < ActionController::Base
 end
 
 BenchmarksController.view_paths = [File.expand_path("./views", __dir__)]
-controller_view = BenchmarksController.new.view_context
 
 Benchmark.ips do |x|
   x.time = 10
   x.warmup = 2
 
-  x.report("component:") { controller_view.render(NameComponent.new(name: "Fox Mulder")) }
-  x.report("partial:") { controller_view.render("partial", name: "Fox Mulder") }
+  x.report("component:") {
+    controller_view = BenchmarksController.new.view_context
+    controller_view.render(NameComponent.new(name: "Fox Mulder"))
+  }
+  x.report("partial:") {
+    controller_view = BenchmarksController.new.view_context
+    controller_view.render("partial", name: "Fox Mulder")
+  }
 
   x.compare!
 end