--- a/test/app/components/form_for_component.html.erb
+++ b/test/app/components/form_for_component.html.erb
@@ -1,0 +1,3 @@
+<%= form_for Post.new, url: "/" do |form| %>
+  <%= render LabelComponent.new(form: form) %>
+<% end %>