--- a/test/app/components/form_with_component.html.erb
+++ b/test/app/components/form_with_component.html.erb
@@ -1,0 +1,3 @@
+<%= form_with Post.new, url: "/" do |form| %>
+  <%= render LabelComponent.new(form: form) %>
+<% end %>