--- a/test/app/components/form_with_component.rb
+++ b/test/app/components/form_with_component.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+class FormWithComponent < ViewComponent::Base
+  def initialize
+  end
+end