--- a/test/app/components/incompatible_form_component.rb
+++ b/test/app/components/incompatible_form_component.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+class IncompatibleFormComponent < ViewComponent::Base
+  def initialize
+  end
+end