--- a/test/app/components/label_component.html.erb
+++ b/test/app/components/label_component.html.erb
@@ -1,0 +1,5 @@
+<div>
+  <%= form.label :published do %>
+    <%= form.check_box :published %>
+  <% end %>
+</div>