--- a/test/app/components/label_component.rb
+++ b/test/app/components/label_component.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class LabelComponent < ViewComponent::Base
+  def initialize(form:)
+    @form = form
+  end
+
+  private
+
+  attr_reader :form
+end