--- a/test/app/models/post.rb
+++ b/test/app/models/post.rb
@@ -3,5 +3,5 @@
 class Post
   include ActiveModel::Model
 
-  attr_accessor :title
+  attr_accessor :title, :published
 end