@BlakeWilliams 2021-06-20
This change isn't great performance wise. I'm hoping to find some time to try optimizing it.

```
Warming up --------------------------------------
          component:     1.594k i/100ms
            partial:   930.000  i/100ms
Calculating -------------------------------------
          component:     15.495k (± 5.2%) i/s -    154.618k in  10.005081s
            partial:      8.729k (± 5.3%) i/s -     87.420k in  10.043416s

Comparison:
          component::    15494.9 i/s
            partial::     8729.3 i/s - 1.78x  (± 0.00) slower
```