@BlakeWilliams 2021-06-23
Not requiring HAML/Slim seemed to increase the benchmark a bit.

```bash
view_component ❯ be rake benchmark
  * development - set it to false
  * test - set it to false (unless you use a tool that preloads your test environment)
  * production - set it to true

Warming up --------------------------------------
          component:     2.491k i/100ms
            partial:   890.000  i/100ms
Calculating -------------------------------------
          component:     25.042k (± 2.4%) i/s -    251.591k in  10.052708s
            partial:      9.043k (± 2.5%) i/s -     90.780k in  10.045240s

Comparison:
          component::    25042.5 i/s
            partial::     9043.0 i/s - 2.77x  (± 0.00) slower
```