@BlakeWilliams 2021-06-23
Looks like the change I made to the benchmark actually decreased the numbers too.

Here's `main` with the `performance/benchmark.rb` changes applied:

```
view_component ❯ be rake benchmark
Warming up --------------------------------------
          component:     3.265k i/100ms
            partial:   987.000  i/100ms
Calculating -------------------------------------
          component:     34.536k (± 3.6%) i/s -    346.090k in  10.035493s
            partial:     10.089k (± 5.2%) i/s -    100.674k in  10.010669s

Comparison:
          component::    34536.1 i/s
            partial::    10088.7 i/s - 3.42x  (± 0.00) slower
```