@BlakeWilliams 2021-06-24
Even more benchmarks:

**Main**
```
view_component ❯ be rake benchmark
Warming up --------------------------------------
          component:    11.677k i/100ms
            partial:     1.359k i/100ms
Calculating -------------------------------------
          component:    105.925k (± 9.6%) i/s -      1.051M in  10.026516s
            partial:     13.237k (± 6.5%) i/s -    131.823k in  10.003570s

Comparison:
          component::   105924.7 i/s
            partial::    13236.9 i/s - 8.00x  (± 0.00) slower
```

**This branch**
```
view_component ❯ be rake benchmark
Warming up --------------------------------------
          component:     5.265k i/100ms
            partial:     1.377k i/100ms
Calculating -------------------------------------
          component:     56.134k (± 4.8%) i/s -    563.355k in  10.061795s
            partial:     13.015k (± 3.8%) i/s -    130.815k in  10.066347s

Comparison:
          component::    56133.6 i/s
            partial::    13015.2 i/s - 4.31x  (± 0.00) slower
```