@BlakeWilliams 2021-09-29
@23tux Hi! There is. This mostly works, but there is an edge case I'm seeing in production applications that I haven't had time to track down.

It would be super helpful to have a failing test case for that scenario or other scenarios production apps might run into. If anyone would like to try to run this locally and write a failing test case, that would help me move forward with this.