@nicolas-brousse 2021-09-29
:wave: @BlakeWilliams

I've the same or similar issue with `capture` used in components, by using some form helpers.
My latest case was with `label` with a block.

I'm busy this week, but I'll try to found time to write some failing test case.

I've tested your branch on our project and solve the issue btw 👍