@percyhanna 2021-10-02
I recently ran into the same problem, or my assumption is that it is the same root problem anyway. I was trying to call some vanilla Rails helpers from my view component with similar results. My pseudo-code looked like this:

```erb
<%= helpers.my_rails_helper do %>
  <h1>Content inside ViewComponent</h1>
<% end %>
```

A simplified version of the `my_rails_helper` helper method looked like:

```ruby
def my_rails_helper
  contents = capture(&block)

  # run conditional logic based on captured contents
  # ...

  # main output
  content_tag(:div, contents, id: "my-id", ...)
end
```

The net result was that the `capture`d content (the `<h1>`) would be rendered before the `<div>` tag that is returned by `my_rails_helper`, something like:

```html
  <h1>Content inside ViewComponent</h1>
<div id="my-id"></div>
```

I tried various approaches to try and fix this, and was able to get the code below to work. It's entirely possible that this is a bad idea, but I wanted to share in case it provided additional insight into the matter. I added these methods to the `ViewComponent` class, which allowed seemingly transparent access to the normal Rails helpers.

```ruby
def respond_to_missing?(method, include_all)
  helpers.respond_to?(method, include_all) || super
end

def method_missing(method, *args, &block)
  if helpers.respond_to?(method)
    helpers.method(method).unbind.bind(self).call(*args, &block)
  else
    super
  end
end
```