@mhw 2021-11-01
I have a failing test case that I think could be caused by the issue you're trying to fix here, added as the commit https://github.com/mhw/view_component/commit/58dacd7fad5672ac42b0f8ac711e85ff2baf2951

I'm trying to bring some partials across to ViewComponent, and this test replicates an issue I ran into with a partial that uses `turbo_frame_tag`. I think it may also be the cause of #1099. The nesting of the elements gets mixed up, as seems to be the case in other failures, but I'm also seeing HTML markup getting quoted in the output. The generated markup from the test case is currently

```
<span>Hello</span>
<div id="container">&lt;/span&gt;
</div>
```

while I think it should be

```
<div id="container">
<span>Hello</span>
</div>
```

That's assuming I've not made a mistake in the use of ViewComponent.