@mhw 2021-12-21
I've been doing a bit more digging on this one and have a branch ([here](https://github.com/mhw/view_component/tree/global-output)) that takes the work here and addresses a couple of additional issues to give something that's working for me in an app where I'm trying to convert a bunch of nested partials in to ViewComponents.

I think the basic approach here is sound: the Rails `capture` helper (and helpers that use it, like `tag`, `form_with`, `form_for`, etc.) rely on being able to temporarily overwrite the `ActionView::Base` `@output_buffer` to redirect the output from compiled templates. As `ViewComponent::Base` doesn't share this `@output_buffer`, evaluating a block from a template in one context within a template in a different one can result in content going to the wrong buffer. So we need to enlist the output buffers of all the view contexts involved in rendering a view.

The one issue I have run into is with Haml: Haml 5 also patches `ActionView::Base` `output_buffer=`, and this doesn't work nicely with the `ActionViewCompatibility` module in this branch. I've taken a quick look at this, but Haml's buffer handling and how it makes its helpers available in Haml templates looks a bit tricky to handle correctly. The good news is that Haml 6 is based on [hamlit](https://github.com/k0kubun/hamlit), and the ActionView::Base patches are gone. In my (brief) tests the combination of the branch linked above and `hamlit` correctly handles cases like passing a `FormBuilder` between view components.